requirements.txt
web3==5.12.2
psycopg2==2.8.6
cryptography==3.2.1
eth-keys==0.3.3
pysha3==1.0.2
rlp==2.0.0a1
json-rpc==1.13.0
confini==0.3.4
sqlalchemy==1.3.20
